requirements.txt
colorlog
nltk
transformers
torch
torchvision
Pillow
tqdm
